% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myclt.R
\name{myclt}
\alias{myclt}
\title{myclt}
\usage{
myclt(n, iter, a = 0, b = 5)
}
\arguments{
\item{n}{number per sample}

\item{iter}{number of samples}

\item{a}{lower bound}

\item{b}{upper bound}
}
\value{
histogram of sum of samples
}
\description{
myclt
}
\examples{
myclt(10,1000)
}
