% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{ntickets}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{number of seats on aircraft}

\item{gamma}{probability of overbooking}

\item{p}{probability that ticketed passenger boards}
}
\value{
optimal number of tickets to sell, as determined by a discrete distribution and the associated normal approximation; two graphs are also returned to visually demonstrate the optimal number of tickets to sell
}
\description{
ntickets
}
\examples{
ntickets(200,.02,.95)
ntickets(225,.025,.985)
}
